drfjwt/requirements.txt
virtualenv
python==3.6.2
Django==1.9.2
mysqlclient
djangorestframework==3.6.4
djangorestframework-jwt==1.8.0
PyMySQL==0.7.11
PyJWT==1.5.3
social-auth-app-django==1.2.0
social-auth-core==1.4.0
python-social-auth==0.3.6
